Split a file into three
